Created vocabulary package and included license headers
