+++ vocabulary/intersection.go
@@ -12,56 +12,64 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package gnostic_vocabulary
 
 import (
 	metrics "github.com/googleapis/gnostic/metrics"
 )
 
-// mapDifference finds the difference between two Vocabularies.
+// mapIntersection finds the intersection between two Vocabularies.
 // This function takes a Vocabulary and checks if the words within
 // the current Vocabulary already exist within the global Vocabulary.
-// If the word exists in both structures it is removed from the
-// Vocabulary structure.
-func mapDifference(v *metrics.Vocabulary) {
+// If the word exists in both structures it is added to a temp Vocabulary
+// which replaces the old Vocabulary.
+func mapIntersection(v *metrics.Vocabulary) {
+	schemastemp := make(map[string]int)
+	operationIDTemp := make(map[string]int)
+	parametersTemp := make(map[string]int)
+	propertiesTemp := make(map[string]int)
 	for _, s := range v.Schemas {
-		_, ok := schemas[s.Word]
+		value, ok := schemas[s.Word]
 		if ok {
-			delete(schemas, s.Word)
+			schemastemp[s.Word] += (value + int(s.Count))
 		}
 	}
 	for _, op := range v.Operations {
-		_, ok := operationID[op.Word]
+		value, ok := operationID[op.Word]
 		if ok {
-			delete(operationID, op.Word)
+			operationIDTemp[op.Word] += (value + int(op.Count))
 		}
 	}
 	for _, param := range v.Parameters {
-		_, ok := parameters[param.Word]
+		value, ok := parameters[param.Word]
 		if ok {
-			delete(parameters, param.Word)
+			parametersTemp[param.Word] += (value + int(param.Count))
 		}
 	}
 	for _, prop := range v.Properties {
-		_, ok := properties[prop.Word]
+		value, ok := properties[prop.Word]
 		if ok {
-			delete(properties, prop.Word)
+			propertiesTemp[prop.Word] += (value + int(prop.Count))
 		}
 	}
+	schemas = schemastemp
+	operationID = operationIDTemp
+	parameters = parametersTemp
+	properties = propertiesTemp
 }
 
-// vocabularyDifference implements the difference operation between multiple Vocabularies.
+// vocabularyIntersection implements the intersection operation between multiple Vocabularies.
 // The function accepts a slice of Vocabularies and returns a single Vocabulary
-// struct which that contains words that were unique to the first Vocabulary in the slice.
-func vocabularyDifference(vocab []*metrics.Vocabulary) *metrics.Vocabulary {
+// struct which that contains words that were found in all of the Vocabularies.
+func VocabularyIntersection(vocabSlices []*metrics.Vocabulary) *metrics.Vocabulary {
 	schemas = make(map[string]int)
 	operationID = make(map[string]int)
 	parameters = make(map[string]int)
 	properties = make(map[string]int)
 
-	unpackageVocabulary(vocab[0])
-	for i := 1; i < len(vocab); i++ {
-		mapDifference(vocab[i])
+	unpackageVocabulary(vocabSlices[0])
+	for i := 1; i < len(vocabSlices); i++ {
+		mapIntersection(vocabSlices[i])
 	}
 
 	v := &metrics.Vocabulary{
